ABM POI en progreso

+++ src/main/java/ar/edu/utn/dds/grupouno/abmc/consultaExterna/dtos/POI_DTO.java
@@ -78,8 +78,8 @@ public class POI_DTO {
 
 	}
 
-	public double getId() {
-		return id;
+	public int getId() {
+		return (int) id;
 	}
 
 	public void setId(Long id) {
